buyer_backend/requirements.txt
langchain
langchain-core
langgraph
langchain-openai
requests
python-dotenv
web3
